Reconciling ImguiLayer und ExampleLayer Vulkan adjustments - V
1. Started using the MSVC maid for leading, trailing, and whatnot redundant spaces.
2. Tracked down the resources leakages, as per the concerns raised by validation layers. Vulkan you are awesome!
3. Anything else I missed

+++ Karma/src/Platform/Vulkan/VulkanBuffer.cpp
@@ -294,7 +294,7 @@ namespace Karma
 		}
 	}
 
-	void VulkanUniformBuffer::UploadUniformBuffer(int frameIndex)
+	void VulkanUniformBuffer::UploadUniformBuffer(size_t frameIndex)
 	{
 		uint32_t index = 0;
 		for (auto& it : GetUniformList())
